pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py2slack"
version = "0.3.2"
description = "Utility functions to send Slack notifications using slack_sdk."
readme = "README.md"
requires-python = ">=3.6"
license = { file = "LICENSE" }
authors = [
    { name = "docto-rin", email = "terriblestomachache@yahoo.co.jp" }
]
dependencies = [
    "slack_sdk",
    "python-dotenv"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[project.urls]
Homepage = "https://github.com/docto-rin/py2slack"
